Patch new ConPTY from Windows, so that image is working in WSL Wezterm
These binaries were released by Windows under MIT License

Prior to this method was launching wezterm with ssh to enable perfect
image viewing experience in wezterm:
- wezterm ssh 127.0.0.1

Now we have this method, we need to copy paste to the destinated Wezterm
installation path in windows. This is the required workarounds for
Wezterm version < Feb 2025.

Tracked upstream:
https://github.com/wezterm/wezterm/tree/main/assets/windows/conhost