feat(nvim): simplify conform.nvim config and add lz alias
Refactored the conform.nvim plugin configuration to remove the use of
function-based options in favor of direct calls to M.resolve_config for
formatter arguments. This streamlines the setup and improves
readability.
This also solved the issue of overriding upstream Lazyvim formatters and
formatters_by_ft
Also added a new 'lz' alias for lazygit in .zsh aliases for convenience.
Additionally, reformatted cspell.json for better readability.